Improve test descriptions
