--- src/main/steps/applicant1/interim-applications/alternative-service/service-application/post.ts
@@ -1,12 +1,0 @@
-import autobind from 'autobind-decorator';
-
-import { InterimApplicationType } from '../../../../../app/case/definition';
-import { AnyObject } from '../../../../../app/controller/PostController';
-import StartInterimApplicationPostController from '../../common/start-interim-application/post';
-
-@autobind
-export default class AlternativeServiceApplicationPostController extends StartInterimApplicationPostController<AnyObject> {
-  protected interimApplicationType(): InterimApplicationType {
-    return InterimApplicationType.ALTERNATIVE_SERVICE;
-  }
-}
