+++ src/main/steps/applicant1/interim-applications/bailiff-service/service-application/post.ts
@@ -5,7 +5,7 @@ import { AnyObject } from '../../../../../app/controller/PostController';
 import StartInterimApplicationPostController from '../../common/start-interim-application/post';
 
 @autobind
-export default class BailiffApplicationPostController extends StartInterimApplicationPostController<AnyObject> {
+export default class BailiffServiceApplicationPostController extends StartInterimApplicationPostController<AnyObject> {
   protected interimApplicationType(): InterimApplicationType {
     return InterimApplicationType.BAILIFF_SERVICE;
   }
